refactor: Auth 도메인 관련 VO 파일 리패키징
